0.2: listener mode, low-latency tuning, Android TV
## Protocol / media
- Listener mode wired end-to-end. JNI nativeConnect now returns the live
  socket handle as a jlong so the accepted-peer socket can replace the
  listener handle on the Kotlin side after srt_accept. Error path returns a
  negative KP_ERR_* sentinel as before.
- SRTO_PEERLATENCY set alongside SRTO_RCVLATENCY. The receiver-side slider
  now actually caps the negotiated TSBPD latency — without PEERLATENCY the
  sender's higher value would win the handshake.
- SRT reject reason is logged on connect failure (e.g. MessageAPI/StreamAPI
  collision, BadSecret).

## Low-latency playback
- Custom LowLatencyRenderersFactory injects MediaFormat.KEY_LOW_LATENCY = 1
  on Android 11+, removing the HW decoder's reorder buffer (~100 ms saved
  on HEVC).
- LoadControl bufferForPlaybackMs lowered to 0 (first-frame-as-soon-as-
  decoded), minBufferMs to 100, bufferForPlaybackAfterRebufferMs to 50.
- Lifecycle resume now always forces a full SRT reconnect rather than only
  on ERROR/IDLE. Backgrounding the app stops accumulating receiver-buffer
  latency on the next foreground.

## Android TV
- Manifest gains uses-feature leanback + touchscreen as `required=false`
  plus a LEANBACK_LAUNCHER intent-filter; the same APK is now eligible for
  both the phone launcher and the Leanback launcher.
- TV banner vector (320×180) with the SMPTE-bars motif.
- armeabi-v7a added to abiFilters so cheap STBs with 32-bit userspace
  (Amlogic S905X on ZTE B866 etc.) install without `INSTALL_FAILED_NO_*`.
  Triples the ABI count for the EP libsrt+mbedtls build but only adds
  ~2 MB to the APK.
- PlayerScreen: BackHandler maps remote BACK to "unlock or disconnect";
  D-pad CENTER toggles the stats overlay. Touch-only swipe gestures are
  skipped on TV where they would be no-ops anyway.

## UI overhaul
- New QuickConnectScreen is the landing on every platform now: shows a
  one-line summary of the saved target (host:port · mode · latency, plus
  passphrase / streamid flags when set) above a big Connect button and a
  Settings entry. Back from Settings always returns here.
- Focus styling: a focused control flips to a solid-white container with
  black content via FocusableButton / FocusableOutlinedButton /
  FocusableIconButton / FocusableFilterChip wrappers. Replaces the earlier
  border-based approach which mis-rendered on TV scaling.
- About dialog now shows the app version (PackageInfo) and a clickable
  mailto link to the author. Bilingual EN+RU thank-you line.
- ConnectionScreen has a Mode chip selector (Caller / Listener /
  Rendezvous), a bind-address hint that shows the local IPv4 in Listener
  mode, manual latency input (text field + ± buttons), and a per-mode
  bandwidth-limit explanation. Saved fields persist across PLAYER →
  SETTINGS transitions inside the same session (not just on cold start).
- QuickConnect carries a low-alpha SMPTE test-pattern background as a
  subtle texture.
- Info icon is sized 28dp and uses the theme primary colour so it reads
  against the dark background.

## Packaging / signing
- Release APK is now named KarPlayer-<versionName>-<buildType>.apk via
  outputFileName overrides.

+++ srt/src/main/kotlin/com/karplayer/srt/SrtNative.kt
@@ -14,6 +14,11 @@ internal object SrtNative {
 
     external fun nativeCreate(): Long
 
+    /**
+     * Returns the SRT socket handle to use for subsequent read/close, or a
+     * negative KP_ERR_* code on failure. In LISTENER mode the returned handle
+     * is the *accepted* peer socket (the listener handle is closed natively).
+     */
     external fun nativeConnect(
         handle: Long,
         host: String,
@@ -26,7 +31,7 @@ internal object SrtNative {
         timeoutMs: Int,
         passphrase: String,
         pbkeylen: Int
-    ): Int
+    ): Long
 
     external fun nativeRead(handle: Long, buffer: ByteArray, offset: Int, length: Int): Int
     external fun nativeClose(handle: Long)
